refactor: P1 TK companion 抽取（功能不变）
将 gateway Models catalog、cyber policy、HandleUpstreamError/model-not-found、
admin available-models/BatchCreate、gateway OpenRouter/Grok voice/composite
抽到 *_tk_* companion；同步 sentinel/agent contract/route audit 扫描面。
no-web-impact
sentinel-registry-reviewed
