+++ backend/internal/handler/admin/account_handler.go
@@ -1802,125 +1802,7 @@ func (h *AccountHandler) BatchCreate(c *gin.Context) {
 	}
 
 	executeAdminIdempotentJSON(c, "admin.accounts.batch_create", req, service.DefaultWriteIdempotencyTTL(), func(ctx context.Context) (any, error) {
-		success := 0
-		failed := 0
-		results := make([]gin.H, 0, len(req.Accounts))
-		// 收集需要异步设置隐私的 OAuth 账号
-		var antigravityPrivacyAccounts []*service.Account
-		var openaiPrivacyAccounts []*service.Account
-
-		for _, item := range req.Accounts {
-			if item.RateMultiplier != nil && *item.RateMultiplier < 0 {
-				failed++
-				results = append(results, gin.H{
-					"name":    item.Name,
-					"success": false,
-					"error":   "rate_multiplier must be >= 0",
-				})
-				continue
-			}
-			// US-024: 单条 Create 走 tkValidateNewAPIAccountCreate，BatchCreate 此前漏调，
-			// 导致 newapi 行只能在 service 层被 "channel_type must be > 0" 拦截，错误信息
-			// 不一致；同时 channel_type / load_factor 之前未透传，使 newapi 批量创建在
-			// service 层 100% 失败。这里补齐验证 + 字段透传。
-			if msg := tkValidateNewAPIAccountCreate(item.Platform, item.ChannelType, item.Credentials); msg != "" {
-				failed++
-				results = append(results, gin.H{
-					"name":    item.Name,
-					"success": false,
-					"error":   msg,
-				})
-				continue
-			}
-
-			// base_rpm 输入校验：负值归零，超过 10000 截断
-			sanitizeExtraBaseRPM(item.Extra)
-
-			skipCheck := item.ConfirmMixedChannelRisk != nil && *item.ConfirmMixedChannelRisk
-
-			account, err := h.adminService.CreateAccount(ctx, &service.CreateAccountInput{
-				Name:                  item.Name,
-				Notes:                 item.Notes,
-				Platform:              item.Platform,
-				Type:                  item.Type,
-				Credentials:           item.Credentials,
-				Extra:                 item.Extra,
-				ProxyID:               item.ProxyID,
-				Concurrency:           item.Concurrency,
-				Priority:              item.Priority,
-				ChannelType:           item.ChannelType,
-				RateMultiplier:        item.RateMultiplier,
-				LoadFactor:            item.LoadFactor,
-				GroupIDs:              item.GroupIDs,
-				ExpiresAt:             item.ExpiresAt,
-				AutoPauseOnExpired:    item.AutoPauseOnExpired,
-				SkipMixedChannelCheck: skipCheck,
-			})
-			if err != nil {
-				failed++
-				results = append(results, gin.H{
-					"name":    item.Name,
-					"success": false,
-					"error":   err.Error(),
-				})
-				continue
-			}
-			// 收集需要异步设置隐私的 OAuth 账号
-			if account.Type == service.AccountTypeOAuth {
-				switch account.Platform {
-				case service.PlatformAntigravity:
-					antigravityPrivacyAccounts = append(antigravityPrivacyAccounts, account)
-				case service.PlatformOpenAI:
-					openaiPrivacyAccounts = append(openaiPrivacyAccounts, account)
-				}
-			}
-			// OpenAI APIKey 账号异步探测 /v1/responses 能力。
-			h.scheduleProtocolCapabilityProbes(account)
-			h.scheduleGrokImportProbe(account)
-			success++
-			results = append(results, gin.H{
-				"name":    item.Name,
-				"id":      account.ID,
-				"success": true,
-			})
-		}
-
-		// 异步设置隐私，避免批量创建时阻塞请求
-		adminSvc := h.adminService
-		if len(antigravityPrivacyAccounts) > 0 {
-			accounts := antigravityPrivacyAccounts
-			go func() {
-				defer func() {
-					if r := recover(); r != nil {
-						slog.Error("batch_create_antigravity_privacy_panic", "recover", r)
-					}
-				}()
-				bgCtx := context.Background()
-				for _, acc := range accounts {
-					adminSvc.ForceAntigravityPrivacy(bgCtx, acc)
-				}
-			}()
-		}
-		if len(openaiPrivacyAccounts) > 0 {
-			accounts := openaiPrivacyAccounts
-			go func() {
-				defer func() {
-					if r := recover(); r != nil {
-						slog.Error("batch_create_openai_privacy_panic", "recover", r)
-					}
-				}()
-				bgCtx := context.Background()
-				for _, acc := range accounts {
-					adminSvc.ForceOpenAIPrivacy(bgCtx, acc)
-				}
-			}()
-		}
-
-		return gin.H{
-			"success": success,
-			"failed":  failed,
-			"results": results,
-		}, nil
+		return h.tkBatchCreateAccounts(ctx, req.Accounts)
 	})
 }
 
@@ -2521,121 +2403,8 @@ func (h *AccountHandler) GetAvailableModels(c *gin.Context) {
 		return
 	}
 
-	// Handle OpenAI accounts
-	if account.IsOpenAI() {
-		// OpenAI 自动透传会绕过常规模型改写，测试/模型列表也应回落到默认模型集。
-		if account.IsOpenAIPassthroughEnabled() {
-			response.Success(c, tkOpenAIAdminDefaultModels(c.Request.Context()))
-			return
-		}
-
-		mapping := account.GetModelMapping()
-		if len(mapping) == 0 {
-			response.Success(c, tkOpenAIAdminDefaultModels(c.Request.Context()))
-			return
-		}
-
-		response.Success(c, tkOpenAIAdminModelsForIDs(sortedModelMappingKeys(mapping)))
-		return
-	}
-
-	// Handle Gemini accounts via runtime model_mapping SSOT (includes Google One defaults).
-	if account.IsGemini() {
-		response.Success(c, tkGeminiAdminAvailableModels(c.Request.Context(), account))
-		return
-	}
-
-	// Handle Antigravity accounts: live servable set (same SSOT as /antigravity/models).
-	if account.Platform == service.PlatformAntigravity {
-		response.Success(c, tkAntigravityAdminDefaultModels(c.Request.Context(), account))
-		return
-	}
-
-	// Handle fifth platform `newapi` accounts.
-	// newapi accounts route OpenAI-compatible payloads through the new-api adaptor
-	// pool, so the model space is whatever the upstream channel exposes. The admin
-	// UI has a dedicated probe (POST /api/v1/admin/channel-types/fetch-upstream-models)
-	// for live model discovery; this endpoint must NOT fall through to the Claude
-	// catalog. We mirror openai's behavior: prefer model_mapping keys when set,
-	// otherwise return an empty list (the UI shows "configure model_mapping" hint).
-	if account.Platform == service.PlatformNewAPI {
-		mapping := account.GetModelMapping()
-		if len(mapping) == 0 {
-			if tkRespondNewAPIAgentPlanAvailableModelsWhenMappingEmpty(c, account) {
-				return
-			}
-			ids, err := h.adminService.GetAccountModelMappingPresetIDs(
-				c.Request.Context(),
-				service.PlatformNewAPI,
-				account.ChannelType,
-			)
-			if err != nil {
-				response.Error(c, http.StatusInternalServerError, "failed to resolve model mapping preset")
-				return
-			}
-			if len(ids) > 0 {
-				sort.Strings(ids)
-				response.Success(c, tkAdminModelOptionsForIDs(ids))
-				return
-			}
-			response.Success(c, []dto.AccountModelOption{})
-			return
-		}
-		response.Success(c, tkAdminModelOptionsForIDs(sortedModelMappingKeys(mapping)))
-		return
-	}
-
-	if account.IsGrok() {
-		if !accountHasExplicitModelMapping(account) {
-			response.Success(c, tkGrokAdminDefaultModels(c.Request.Context()))
-			return
-		}
-		mapping := account.GetModelMapping()
-		if len(mapping) == 0 {
-			response.Success(c, tkGrokAdminDefaultModels(c.Request.Context()))
-			return
-		}
-		// Deterministic order: pin the grok chat-probe default first, then
-		// alphabetical. Iterating the mapping map directly is non-deterministic
-		// and makes the admin selector (and its test) flaky.
-		ids := make([]string, 0, len(mapping))
-		for requestedModel := range mapping {
-			ids = append(ids, requestedModel)
-		}
-		sort.SliceStable(ids, func(i, j int) bool {
-			if ids[i] == service.GrokDefaultTestModelID {
-				return true
-			}
-			if ids[j] == service.GrokDefaultTestModelID {
-				return false
-			}
-			return ids[i] < ids[j]
-		})
-		response.Success(c, tkGrokAdminModelsForIDs(ids))
-		return
-	}
-
-	if account.IsKiro() || account.IsKiroMirrorStub() {
-		response.Success(c, tkClaudeModelsToAdminOptions(service.KiroAdminTestModels(), nil))
-		return
-	}
-
-	// Handle Claude/Anthropic accounts
-	// For OAuth and Setup-Token accounts: return default models
-	if account.IsOAuth() {
-		response.Success(c, tkClaudeAdminDefaultModels(c.Request.Context()))
-		return
-	}
-
-	// For API Key accounts: return models based on model_mapping
-	mapping := account.GetModelMapping()
-	if len(mapping) == 0 {
-		// No mapping configured, return default models
-		response.Success(c, tkClaudeAdminDefaultModels(c.Request.Context()))
-		return
-	}
-
-	response.Success(c, tkClaudeAdminModelsForIDs(sortedModelMappingKeys(mapping)))
+	// TK: platform model catalog dispatch — see account_handler_tk_available_models.go
+	h.tkRespondAvailableModels(c, account)
 }
 
 // SyncUpstreamModels handles syncing live supported models from an account's upstream.
