+++ backend/internal/service/ratelimit_service.go
@@ -14,7 +14,6 @@ import (
 	"time"
 
 	"github.com/Wei-Shaw/sub2api/internal/config"
-	"github.com/Wei-Shaw/sub2api/internal/pkg/logger"
 	"github.com/tidwall/gjson"
 )
 
@@ -377,191 +376,10 @@ func (s *RateLimitService) HandleUpstreamError(ctx context.Context, account *Acc
 	// 同请求内与 fastpath 调用点的重复触发由方法内去重吸收。
 	s.maybeHandleOpenAITeamLinkedError(ctx, account, statusCode, responseBody)
 	customErrorCodesEnabled := account.IsCustomErrorCodesEnabled()
-
 	if handled, disable := s.tkTryHandleUpstreamErrorPrelude(ctx, account, statusCode, headers, responseBody, firstRequestedModel(requestedModel)); handled {
 		return disable
 	}
-	// Recompute for pool-mode / custom-error carve-outs (predicate is cheap;
-	// prelude already applied the CloudWise cooldown early returns above).
-	cloudwiseModelBalance402 := tkIsCloudwiseModelBalance402Response(account, statusCode, responseBody)
-
-	// 池模式默认不标记本地账号状态；但管理员显式配置的临时不可调度规则优先。
-	// 401 保留现有认证错误语义，不在这里改变池模式的认证处理。
-	if account.IsPoolMode() && !customErrorCodesEnabled && account.Platform != PlatformAnthropic && !cloudwiseModelBalance402 {
-		if statusCode != http.StatusUnauthorized && s.tryTempUnschedulable(ctx, account, statusCode, responseBody) {
-			return true
-		}
-		slog.Info("pool_mode_error_skipped", "account_id", account.ID, "status_code", statusCode)
-		return false
-	}
-
-	// apikey 类型账号：检查自定义错误码配置
-	// 如果启用且错误码不在列表中，则不处理（不停止调度、不标记限流/过载）
-	planGatedModel := isOpenAIOAuthAccount(account) && isOpenAICodexPlanGatedModelError(statusCode, responseBody)
-	if !cloudwiseModelBalance402 && !planGatedModel && !account.ShouldHandleErrorCode(statusCode) && account.Platform != PlatformAnthropic {
-		slog.Info("account_error_code_skipped", "account_id", account.ID, "status_code", statusCode)
-		return false
-	}
-
-	if len(requestedModel) > 0 && s.HandleUpstreamModelNotFound(ctx, account, requestedModel[0], statusCode, responseBody) {
-		return true
-	}
-
-	// Anthropic official 5h / 7d window exhaustion is a hard account limit.
-	// It must take precedence over user-configured 429 temp-unsched rules,
-	// otherwise a broad "rate limit" keyword rule can shorten a multi-hour
-	// cooldown to a local temporary pause.
-	if statusCode == http.StatusTooManyRequests && account.Platform == PlatformAnthropic {
-		// 7d_oi 是 Fable 模型专属的 7d 窗口：只标记模型级限流，账号对其他模型仍可调度。
-		fableLimited := s.persistAnthropicFableWindowLimit(ctx, account, headers)
-		if s.persistAnthropicExhaustedWindowLimit(ctx, account, headers) {
-			return false
-		}
-		if fableLimited {
-			return false
-		}
-	}
-
-	// 先尝试临时不可调度规则（401除外）
-	// 如果匹配成功，直接返回，不执行后续禁用逻辑
-	// 529 overload 走 handle529 全局冷却，优先于账号级 temp-unsched 规则。
-	if statusCode != 401 && statusCode != 529 {
-		if s.tryTempUnschedulable(ctx, account, statusCode, responseBody, firstRequestedModel(requestedModel)) {
-			return true
-		}
-	}
-
-	upstreamMsg := strings.TrimSpace(extractUpstreamErrorMessage(responseBody))
-	upstreamMsg = sanitizeUpstreamErrorMessage(upstreamMsg)
-	if upstreamMsg != "" {
-		upstreamMsg = truncateForLog([]byte(upstreamMsg), 512)
-	}
-
-	if account.Platform == PlatformAnthropic && tkIsAnthropicUsagePolicyBlock(upstreamMsg, responseBody) {
-		return s.handleAnthropicUsagePolicyBlock(ctx, account, statusCode, upstreamMsg, responseBody)
-	}
-
-	if handled, quotaDisable := s.tkMaybeHandleKiroEndpointQuotaExhausted(ctx, account, upstreamMsg, responseBody); handled {
-		return quotaDisable
-	}
-
-	switch statusCode {
-	case 400:
-		// "organization has been disabled" → 永久禁用
-		if strings.Contains(strings.ToLower(upstreamMsg), "organization has been disabled") {
-			msg := "Organization disabled (400): " + upstreamMsg
-			s.handleAuthError(ctx, account, msg)
-			shouldDisable = true
-		} else if account.Platform == PlatformAnthropic && strings.Contains(strings.ToLower(upstreamMsg), "credit balance") {
-			// Anthropic API key 余额不足（语义等同 402），停止调度
-			msg := "Credit balance exhausted (400): " + upstreamMsg
-			s.handleAuthError(ctx, account, msg)
-			shouldDisable = true
-		} else if strings.Contains(strings.ToLower(upstreamMsg), "identity verification is required") {
-			// KYC 身份验证要求 → 永久禁用，账号需完成身份验证后才能恢复
-			msg := "Identity verification required (400): " + upstreamMsg
-			s.handleAuthError(ctx, account, msg)
-			shouldDisable = true
-		} else if account.Platform == PlatformOpenAI && isOpenAIImageCapabilityLoss400(statusCode, responseBody) {
-			_ = s.HandleOpenAIImageCapabilityLoss400(ctx, account, statusCode, responseBody)
-		} else if account.Platform == PlatformAnthropic && tkIsAnthropicClientInducedBadRequest(responseBody) {
-			slog.Info("anthropic_client_induced_400_skip_penalty",
-				"account_id", account.ID, "status_code", statusCode)
-		}
-		// 其他 400 错误（如参数问题）不处理，不禁用账号
-	case 401:
-		shouldDisable = s.tkHandleAuth401(ctx, account, upstreamMsg, responseBody)
-	case 402:
-		// 国产供应商：余额不足是可恢复状态（充值/检测恢复后由周期任务自动解除），
-		// 不能走 handleAuthError 永久置 status=error。改为可恢复的临时停调。
-		if account.IsCNProvider() {
-			s.handleCNProviderInsufficientBalance(ctx, account, upstreamMsg)
-			shouldDisable = true
-			break
-		}
-		// OpenAI: deactivated_workspace 表示工作区已停用，直接标记 error
-		if account.Platform == PlatformOpenAI && gjson.GetBytes(responseBody, "detail.code").String() == "deactivated_workspace" {
-			msg := "Workspace deactivated (402): workspace has been deactivated"
-			s.handleAuthError(ctx, account, msg)
-			shouldDisable = true
-			break
-		}
-		if s.tkHandleKiroQuotaLimit402(ctx, account, upstreamMsg, responseBody) {
-			shouldDisable = true
-			break
-		}
-		// 支付要求：余额不足或计费问题，停止调度
-		msg := "Payment required (402): insufficient balance or billing issue"
-		if upstreamMsg != "" {
-			msg = "Payment required (402): " + upstreamMsg
-		}
-		s.handleAuthError(ctx, account, msg)
-		shouldDisable = true
-	case 403:
-		logger.LegacyPrintf(
-			"service.ratelimit",
-			"[HandleUpstreamErrorRaw] account_id=%d platform=%s type=%s status=403 request_id=%s cf_ray=%s upstream_msg=%s raw_body=%s",
-			account.ID,
-			account.Platform,
-			account.Type,
-			strings.TrimSpace(headers.Get("x-request-id")),
-			strings.TrimSpace(headers.Get("cf-ray")),
-			upstreamMsg,
-			truncateForLog(responseBody, 1024),
-		)
-		shouldDisable = s.handle403(ctx, account, upstreamMsg, responseBody)
-	case 404:
-		shouldDisable = s.handle404(ctx, account, upstreamMsg, responseBody)
-	case 413:
-		if account.Platform == PlatformAnthropic {
-			slog.Info("anthropic_request_too_large_skip_penalty",
-				"account_id", account.ID,
-				"status_code", statusCode)
-		}
-	case 429:
-		if handled, disable := s.tkTryHandle429Extras(ctx, account, statusCode, headers, upstreamMsg, responseBody); handled {
-			return disable
-		}
-		rateLimitSet := s.handle429(ctx, account, headers, responseBody, requestedModel...)
-		if account.Platform == PlatformAnthropic && tkAnthropicStubHealthFuseEligible(statusCode, rateLimitSet) {
-			shouldDisable = s.handleAnthropicUpstreamErrorWithOptions(ctx, account, statusCode, upstreamMsg, responseBody, rateLimitSet)
-		} else {
-			shouldDisable = false
-		}
-	case 529:
-		if customErrorCodesEnabled && account.ShouldHandleErrorCode(statusCode) {
-			msg := "Custom error code triggered"
-			if upstreamMsg != "" {
-				msg = upstreamMsg
-			}
-			s.handleCustomErrorCode(ctx, account, statusCode, msg)
-			shouldDisable = true
-		} else {
-			retryAfterOwned := s.handle529(ctx, account, headers)
-			if account.Platform == PlatformAnthropic {
-				shouldDisable = s.handleAnthropicUpstreamErrorWithOptions(ctx, account, statusCode, upstreamMsg, responseBody, retryAfterOwned)
-			} else {
-				shouldDisable = false
-			}
-		}
-	default:
-		if account.Platform == PlatformAnthropic && tkAnthropicStubHealthFuseEligible(statusCode, false) {
-			shouldDisable = s.handleAnthropicUpstreamError(ctx, account, statusCode, upstreamMsg, responseBody)
-		} else if customErrorCodesEnabled {
-			msg := "Custom error code triggered"
-			if upstreamMsg != "" {
-				msg = upstreamMsg
-			}
-			s.handleCustomErrorCode(ctx, account, statusCode, msg)
-			shouldDisable = true
-		} else if statusCode >= 500 {
-			// 未启用自定义错误码时：仅记录5xx错误
-			slog.Warn("account_upstream_error", "account_id", account.ID, "status_code", statusCode)
-			shouldDisable = false
-		}
-	}
-
-	return shouldDisable
+	return s.tkDispatchHandleUpstreamError(ctx, account, statusCode, headers, responseBody, requestedModel, customErrorCodesEnabled)
 }
 
 // PreCheckUsage proactively checks local quota before dispatching a request.
@@ -2301,104 +2119,8 @@ func parseOpenAIImageTryAgainCooldown(body []byte) time.Duration {
 	}
 }
 
-const upstreamModelNotFoundCooldown = 30 * time.Minute
-const upstreamModelNotFoundReason = "upstream_404_model_not_found"
-const upstreamCodexPlanGatedModelCooldown = 30 * time.Minute
-const upstreamCodexPlanGatedModelReason = "upstream_400_codex_plan_gated_model"
 const tempUnschedBodyMaxBytes = 64 << 10
 const tempUnschedMessageMaxBytes = 2048
-
-// HandleUpstreamModelNotFound marks the requested model as temporarily
-// unavailable on the account when the upstream deterministically reports it
-// cannot serve that model: a 404 model-not-found, or the Codex 400 rejecting a
-// plan-gated model on a ChatGPT OAuth account. Returning true tells the caller
-// to fail the current attempt over to another account; the scheduler skips the
-// (account, model) pair via IsSchedulableForModelWithContext until the
-// cooldown expires, instead of re-selecting an account that can never serve
-// the model.
-func (s *RateLimitService) HandleUpstreamModelNotFound(ctx context.Context, account *Account, requestedModel string, statusCode int, responseBody []byte) bool {
-	if s == nil || account == nil || s.accountRepo == nil {
-		return false
-	}
-	if !account.ShouldHandleErrorCode(statusCode) {
-		return false
-	}
-	var cooldown time.Duration
-	var reason string
-	switch {
-	case isUpstreamModelNotFoundError(statusCode, responseBody):
-		cooldown, reason = upstreamModelNotFoundCooldown, upstreamModelNotFoundReason
-	case isOpenAIOAuthAccount(account) && isOpenAICodexPlanGatedModelError(statusCode, responseBody):
-		cooldown, reason = upstreamCodexPlanGatedModelCooldown, upstreamCodexPlanGatedModelReason
-	default:
-		return false
-	}
-	if account.Platform == PlatformAnthropic {
-		slog.Info("anthropic_model_not_found_skip_penalty",
-			"account_id", account.ID,
-			"requested_model", strings.TrimSpace(requestedModel))
-		return false
-	}
-	modelKey := modelRateLimitKeyForUpstreamModelNotFound(ctx, account, requestedModel)
-	if modelKey == "" {
-		return false
-	}
-	if reason == upstreamModelNotFoundReason && account.Platform == PlatformOpenAI && account.IsOAuth() && account.IsModelSupported(requestedModel) {
-		if !IsGPTImageGenerationModel(requestedModel) && !IsGPTImageGenerationModel(modelKey) {
-			return false
-		}
-	}
-	if shouldSkipCodexPlanGatedImageModelCooldown(ctx, reason, requestedModel, modelKey) {
-		return true
-	}
-	resetAt := time.Now().Add(cooldown)
-	if err := s.accountRepo.SetModelRateLimit(ctx, account.ID, modelKey, resetAt, reason); err != nil {
-		slog.Warn("upstream_model_not_found_set_model_rate_limit_failed", "account_id", account.ID, "model", modelKey, "reason", reason, "error", err)
-		return true
-	}
-	slog.Info("upstream_model_not_found_model_rate_limited", "account_id", account.ID, "model", modelKey, "reason", reason, "reset_at", resetAt)
-	return true
-}
-
-// shouldSkipCodexPlanGatedImageModelCooldown 判断这次 Codex plan-gated 400 是否
-// 属于"图片模型被文本端点拒绝"。
-//
-// 这类错误是确定性的端点错配，不是账号能力缺失：同一账号通过 /v1/images/* 依然
-// 能出图。在这里写 per-model 冷却，会让一次用错端点的请求把整个号池对正确的生图
-// 端点下线（#4828）。
-//
-// 但请求本身就从 /v1/images/* 入站时不适用——那种情况下被拒说明账号确实不具备
-// 该模型能力，冷却是必要的刹车：没有它，每个生图请求都会完整走一遍号池，对上游
-// 形成无上界的 400 放大。
-//
-// 请求模型与最终冷却键都要判：冷却键走的是 account.GetMappedModel，账号可能把
-// 文本别名映射到 gpt-image-*，只判请求模型会漏掉这种形态。
-func shouldSkipCodexPlanGatedImageModelCooldown(ctx context.Context, reason, requestedModel, modelKey string) bool {
-	if reason != upstreamCodexPlanGatedModelReason {
-		return false
-	}
-	if OpenAIImagesEndpointFromContext(ctx) {
-		return false
-	}
-	return IsGPTImageGenerationModel(requestedModel) || IsGPTImageGenerationModel(modelKey)
-}
-
-func modelRateLimitKeyForUpstreamModelNotFound(ctx context.Context, account *Account, requestedModel string) string {
-	modelKey := strings.TrimSpace(requestedModel)
-	if account == nil || modelKey == "" {
-		return modelKey
-	}
-	if account.Platform == PlatformAntigravity {
-		if resolved := strings.TrimSpace(resolveFinalAntigravityModelKey(ctx, account, modelKey)); resolved != "" {
-			return resolved
-		}
-		return modelKey
-	}
-	if mapped := strings.TrimSpace(account.GetMappedModel(modelKey)); mapped != "" {
-		return mapped
-	}
-	return modelKey
-}
 
 func firstRequestedModel(requestedModel []string) string {
 	if len(requestedModel) == 0 {
