+++ scripts/sentinels/pricing-availability.json
@@ -6,8 +6,6 @@
       "path": "backend/internal/handler/gateway_handler.go",
       "must_contain": [
         "TkRecordFailureFromErr",
-        "tkFilterModelIDs",
-        "tkGeminiDefaultModelsList",
         "TkEvalBodyGuard("
       ],
       "rationale": "gateway_handler.go has TK injection points across multiple flows: (1) failure taps at the Anthropic /v1/messages + Gemini /v1/messages error paths (TkRecordFailureFromErr), (2) the CatalogPolicy model-list projection on the whitelist path (tkFilterModelIDs), (3) the gemini /v1/models fallback re-point to the CatalogPolicy projection (tkGeminiDefaultModelsList) so the list converges with /pricing instead of returning the raw geminicli.DefaultModels (which still carried advertised_dead gemini-2.0-flash) — an upstream merge could silently revert this one-line branch back to the raw list and un-converge gemini /v1/models, and (4) the pre-flight body-size guard on Messages + CountTokens entries (TkEvalBodyGuard) that fail-fast oversize requests before forward — see edge:us1 production data where opus-4-7 upstream rejects body >= ~940 KB with raw 403. Antigravity default-model projection and 403 enrichment live in TK companions (gateway_handler_tk_model_list.go / gateway_handler_tk_forward_error.go)."
@@ -473,11 +471,10 @@
     {
       "path": "backend/internal/handler/admin/account_handler.go",
       "must_contain": [
-        "tkGeminiAdminAvailableModels(c.Request.Context(), account)",
-        "tkClaudeAdminModelsForIDs(sortedModelMappingKeys(mapping))",
+        "h.tkRespondAvailableModels(c, account)",
         "scheduleProtocolCapabilityProbes"
       ],
-      "rationale": "Admin GetAvailableModels dispatcher keeps the Gemini/Claude menu call sites and create/update/re-test scheduleProtocolCapabilityProbes injection points after TK companion extraction. Model builders and protocol-probe implementation live in account_handler_tk_available_models.go / account_handler_tk_protocol_probe.go."
+      "rationale": "Admin GetAvailableModels stays a thin dispatcher calling tkRespondAvailableModels; create/update/re-test still inject scheduleProtocolCapabilityProbes. Model builders live in account_handler_tk_available_models.go / protocol-probe companion."
     },
     {
       "path": "backend/internal/handler/admin/account_handler_tk_available_models.go",
@@ -486,7 +483,10 @@
         "func tkGeminiAdminModelsForMapping(",
         "service.ServableClientFacingIDs(ctx, service.PlatformGemini, nil, nil)",
         "func tkAdminModelOptions[",
-        "[]dto.AccountModelOption"
+        "[]dto.AccountModelOption",
+        "func (h *AccountHandler) tkRespondAvailableModels(",
+        "tkGeminiAdminAvailableModels(c.Request.Context(), account)",
+        "tkClaudeAdminModelsForIDs(sortedModelMappingKeys(mapping))"
       ],
       "rationale": "Admin available-models TK builders: Gemini routes through tkGeminiAdminAvailableModels (model_mapping SSOT). Non-Google-One accounts still intersect mapping with ServableClientFacingIDs via tkGeminiAdminModelsForMapping; Google One keeps the conservative whitelist from GetModelMapping without servable pruning. Every platform projects package-specific metadata through tkAdminModelOptions into the same []dto.AccountModelOption HTTP contract."
     },
@@ -527,6 +527,14 @@
         "tkModelPricingFromLiteLLM(loadTKPricingOverlay()[owner])"
       ],
       "rationale": "BillingService pricing hooks: unknown zeroes fail closed, interval/thinking/media fields survive conversion, Imagen keeps flat billing, and compatibility aliases normalize. Production alias matching must resolve an active registry owner instead of using the legacy numeric Go fixture table. Not every registry row is a family alias; dropping the scoped hooks reintroduces price drift while broadening them changes fail-closed behavior for unrelated models. (LiteLLM/overlay conversion moved to companion)"
+    },
+    {
+      "path": "backend/internal/handler/gateway_handler_tk_models_catalog.go",
+      "must_contain": [
+        "tkFilterModelIDs",
+        "tkGeminiDefaultModelsList"
+      ],
+      "rationale": "gateway_handler.go has TK injection points across multiple flows: (1) failure taps at the Anthropic /v1/messages + Gemini /v1/messages error paths (TkRecordFailureFromErr), (2) the CatalogPolicy model-list projection on the whitelist path (tkFilterModelIDs), (3) the gemini /v1/models fallback re-point to the CatalogPolicy projection (tkGeminiDefaultModelsList) so the list converges with /pricing instead of returning the raw geminicli.DefaultModels (which still carried advertised_dead gemini-2.0-"
     }
   ]
 }
